Start introducing domain layer
Add namespaces for the major domain concepts, with schema, coercion, and other
access functions. These are all pure-data cljc, it is up to the caller to hook
these up to APIs, storage, etc.

First version of the workspace-store API, which is being leveraged in
slack-widgets to keep a mutable store of the workspace state.